==> beyond_looks_analyses/1.1_clean_KBH_archival_data.Rmd <==
---
title: "1_clean_data"
author: "ASM"
date: "2024-01-12"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#1. Load libraries
```{r}
library(tidyverse)
library(readxl)
library(broom)
library(readr)
library(janitor)
library(here)
library(readr)
library(dplyr)
library(data.table)
library(tidylog)
library(stringr)
```

#Load and Clean data
## A. Demographic data
```{r}
#This section reads in and cleans the variable names of the demographic data for all the studies.
#It only needs to be loaded once and then the merged demographics can be loaded as a single RDA file

load (here("archival_KBH_anonimyzed/anonymized_demographic_data/demo_learn.Rda")) 
  # N=30
  
load(here("archival_KBH_anonimyzed/anonymized_demographic_data/demo_cogmisp.Rda"))
  # N= 76
  
load(here("archival_KBH_anonimyzed/anonymized_demographic_data/demo_mix14.Rda")) 
  # N=49
  
load(here("archival_KBH_anonimyzed/anonymized_demographic_data/demo_mix20.Rda")) 
  # N=67
  
load(here("archival_KBH_anonimyzed/anonymized_demographic_data/demo_comp.Rda")) 
  # N=30
  
  # #All the demographic information gets combined into a single data frame

Demo_arch <- rbind(demo_learn,demo_cogmisp,demo_comp,demo_mix14,demo_mix20) #n=252

save(Demo_arch, file = here("archival_KBH_anonimyzed/anonymized_demographic_data/Demo_arch.Rda"))


```


## B. Eyetracking data
### Columns to keep
```{r}
#In this section, raw eye-tracking data from 5 studies is loaded, cleaned, saved into RDA objects
#It only needs to be loaded once and then the merged demographics can be loaded as a single RDA file


#Specify which columns to keep from many columns exported from TOBII
keepcols <- c("RecordingName","MediaName", "RecordingDate", "LocalTimeStamp",
            "RecordingTimestamp","ValidityLeft","ValidityRight","GazePointX (ADCSpx)",
            "GazePointY (ADCSpx)", "StudioProjectName", "StudioTestName", "PupilLeft", "PupilRight", "KeyPressEvent", "StudioEvent") 

```

### Read and select columns from first dataset codename "Learn"
```{r}
load(here("archival_KBH_anonimyzed/anonymized_eye_data/eye_data_learn.Rda"))

eye_data_learn <- eye_data_learn %>% 
  select(all_of(keepcols))
```
### Read and select columns from second dataset codename "Cogmisp"
```{r}
load(here("archival_KBH_anonimyzed/anonymized_eye_data/eye_data_cogmisp.Rda"))

eye_data_cogmisp <- eye_data_cogmisp %>%
  select(all_of(keepcols))
```

### Read and select columns from third and fourth datasets codenames "Mix14" "Mix20"
#The eyedata from this studies was separated into multiple datasets
```{r}
#Mix 1
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix1.Rda"))

mix1<- mix1 %>%
  mutate(StudioTestName= as.character(StudioTestName))

#Mix2
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix2.Rda"))

mix2<- mix2 %>%
  mutate(StudioTestName= as.character(StudioTestName))

#Mix3
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix3.Rda"))

mix3<- mix3 %>%
  mutate(StudioTestName= as.character(StudioTestName))

#Mix4
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix4.Rda"))

mix4<- mix4 %>%
  mutate(StudioTestName= as.character(StudioTestName))

#Mix5
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix5.Rda"))

mix5<- mix5 %>%
  mutate(StudioTestName= as.character(StudioTestName))

#Mix6
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix6.Rda"))

mix6<- mix6 %>%
  mutate(StudioTestName= as.character(StudioTestName))

#Mix7
load(here("archival_KBH_anonimyzed/anonymized_eye_data/mix7.Rda"))

mix7<- mix7 %>%
  mutate(StudioTestName= as.character(StudioTestName))

eye_data_mix_studies <- mix1 %>%
  bind_rows(mix2) %>%
  bind_rows(mix3) %>%
  bind_rows(mix4) %>%
  bind_rows(mix5) %>%
  bind_rows(mix6) %>%
  bind_rows(mix7) %>%
  filter(str_detect(RecordingName, "Mix_14")| str_detect(RecordingName, "Mix_20")) %>%
  select(all_of(keepcols)) 

#Data check revealed several problems. Problems and solutions detailed below:

#Problem 1: the studio project name column contained information from project name and test name together
#and all the names were Mix 20, even if they were part of the Mix 14 project.
#Fix:

eye_data_mix_studies <- eye_data_mix_studies %>%
  select(-StudioTestName) %>%
  mutate(StudioTestName = case_when(StudioProjectName =="Mix-20-E1" ~ "E1",
                                    StudioProjectName =="Mix-20-E1 (DO NOT USE)" ~ "E1",
                                    StudioProjectName =="Mix-20-E2" ~ "E2",
                                    StudioProjectName =="Mix-20-E2 (DO NOT USE)" ~ "E2",
                                    StudioProjectName =="Mix-20-F1" ~ "F1",
                                    StudioProjectName =="Mix-20-F1 (DO NOT USE)" ~ "F1",
                                    StudioProjectName =="Mix-20-F2" ~ "F2",
                                    StudioProjectName =="Mix-20-F2 (DO NOT USE)" ~ "F2"))%>%
  select(-StudioProjectName) %>%
  mutate(StudioProjectName = case_when(str_detect(RecordingName, "Mix_20")==TRUE ~ "Mix_20", 
                                       str_detect(RecordingName, "Mix_14")==TRUE ~ "Mix_14"))

#Note that some files were named Do Not Use. They contained subjects that were excluded from the original experiment
#for diverse reasons, however we included all the subjects in our data frame and then specified reasons for exclusions
#e.g. equipment failure in the demographic datasets. This way we kept the exclusion transparent and consistent across data sets.

#Problem 2: the way the auditory stimuli were named was inconsistent
#For example, sometimes the same audio "look at the spoon" was named "Spoon R(2).AVI" and others "Spoon R.AVI".
#sometimes the naming was not inconsistent but just uninformative e.g. "LF1(2).AVI"
#Fix:

eye_data_mix_studies <- eye_data_mix_studies %>%
  mutate(MediaName = case_when(MediaName == "LF1(2).AVI" ~ "Chien L FR.AVI",
                              MediaName == "LF2(2).AVI" ~ "Bouche R SW.AVI",
                              MediaName == "LF3(2).AVI" ~ "Biscuit L FR.AVI",
                              MediaName == "LF4(2).AVI" ~ "Pomme R.AVI",
                              MediaName == "LF5(2).AVI" ~ "Cuillere L.AVI",
                              MediaName == "LF6(2).AVI" ~ "Pied L SW.AVI",
                              MediaName == "LF7(2).AVI" ~ "Livre R FR.AVI",
                              MediaName == "LF8(2).AVI" ~ "Main R.AVI",
                              MediaName == "LF9(2).AVI" ~ "Oreille L.AVI",
                              MediaName == "LF10(2).AVI" ~ "Porte L SW.AVI",
                              MediaName == "LF11(2).AVI" ~ "Biscuit R SW.AVI",
                              MediaName == "LF11.AVI" ~ "Biscuit R SW.AVI",
                              MediaName == "LF12(2).AVI" ~ "Bouche L FR.AVI",
                              MediaName == "LianeF(2).AVI" ~ "Main L.AVI",
                              MediaName == "LF14(2).AVI" ~ "Chien R SW.AVI",
                              MediaName == "LF15(2).AVI" ~ "Pied R FR.AVI",
                              MediaName == "LF16(2).AVI" ~ "Brosse a dent L.AVI",
                              MediaName == "LF17(2).AVI" ~ "Livre R FR.AVI",
                              MediaName == "LF18(2).AVI" ~ "Livre L SW.AVI",
                              MediaName == "Spoon R(2).AVI" ~ "Spoon R.AVI",
                              MediaName == "Foot L SW(2).AVI" ~ "Foot L SW.AVI",
                              MediaName == "Book R ENG(2).AVI" ~ "Book R ENG.AVI",
                              MediaName == "Hand R(8).AVI" ~ "Hand R.AVI",
                              MediaName == "Ear L(2).AVI" ~ "Ear L.AVI",
                              MediaName == "Cookie R SW(2).AVI" ~ "Cookie R SW.AVI",
                              MediaName == "Mouth L ENG(2).AVI" ~ "Mouth L ENG.AVI",
                              MediaName == "Bird L(3).AVI" ~ "Bird L.AVI",
                              MediaName == "Bird R(2).AVI" ~ "Bird R.AVI",
                              MediaName == "LE2.AVI" ~ "Bird R.AVI",
                              MediaName == "LE2(3).AVI" ~ "Bird R.AVI",
                              MediaName == "Dog R SW(3).AVI" ~ "Dog R SW.AVI",
                              MediaName == "Foot R ENG(2).AVI" ~ "Foot R ENG.AVI",
                              MediaName == "Toothbrush L(2).AVI" ~ "Toothbrush L.AVI",
                              MediaName == "Door R ENG(2).AVI" ~ "Door R ENG.AVI",
                              MediaName == "Book L SW(2).AVI" ~ "Book L SW.AVI",
                              MediaName == "Hand L(2).AVI" ~ "Hand L.AVI",
                              MediaName == "Mouth R SW(2).AVI" ~ "Mouth R SW.AVI",
                              MediaName == "Mouth, R, SW.AVI"  ~ "Mouth R SW.AVI",
                              TRUE ~ MediaName))

#Problem 3: 
#Fix:
#wrong recording name for a twin. They had the same recording name but different study order, so they were two separate kids.
eye_data_mix_studies<- eye_data_mix_studies %>%
  mutate(RecordingName = case_when(RecordingName == "Mix_20_S50" & StudioTestName == "E1" ~ "Mix_20_S49", 
                                   TRUE ~ RecordingName))

#Mix20 Data was missing and was non existent in Tobii for subjects: S9, s10, s11, s12, s15 and s46. 


#Mix 14 data for subjects 6 and 3 missing from tobii


eye_data_mix_clean <- eye_data_mix_studies

#save?

```

### Read and select columns from second dataset codename "Comp"
```{r}
load(here("archival_KBH_anonimyzed/anonymized_eye_data/eye_data_comp.Rda"))

eye_data_comp <- eye_data_comp %>%
  select(all_of(keepcols))
```

###Data Merging
#This code merges all the eye gaze data into a single data set and saves the completely new data set 
```{r}
eye_data_arch<- rbind(eye_data_comp,eye_data_mix_clean,eye_data_learn,eye_data_cogmisp)
eye_data_arch <- clean_names(eye_data_arch) %>%
  mutate(studio_project_name = str_replace(studio_project_name, "_", "-")) #change Mix names to have hyphen instead of underscore so they merge correctly later
#save(eye_data_arch, file = here("R Analyses/Eye_data/eye_data_arch2.Rda"))

# #Check potential issues with the merge
# setdiff(Demo_arch$recording_name, eye_data_arch$recording_name) # all the rows that are different correspond to data that has already been identified as missing from tobii
# setdiff(eye_data_arch$recording_name, Demo_arch$recording_name) # 0 so all good!


#Remove attention getters and blank media name rows
arch_comp_data <- eye_data_arch %>%
  filter(str_detect(media_name, "attention")==FALSE & 
           str_detect(media_name, "Getter")==FALSE & 
           str_detect(media_name, "getter")==FALSE & 
           str_detect(media_name, "Attention") == FALSE) %>%
  filter(!is.na(media_name)) %>%
  group_by(recording_name) %>%
  mutate(trial_number = rleid(media_name)) %>% #create trial number
  full_join(Demo_arch) %>% #11 unmatched rows as identified with setdiff above (all identified as missing from Tobii)
  ungroup()

#Adding an Area of Interest column to the newly merged data frame

#This reads in a previously created dataframe containing AOI information for each trial/order/study
AOI<- read_excel(here("archival_KBH_anonimyzed/AOI_maker.xlsx")) %>%
  clean_names() %>%
  mutate(trial_number = as.integer(parse_number(trial_number)))

#Check potential issues with the merging

# setdiff(AOI$media_name, arch_comp_data$media_name) #0 no differences
# setdiff(arch_comp_data$media_name, AOI$media_name)# The differences are just from test trials from the "learn" experiment, for which we are only interested in the learning phase. So no problem.

arch_aoi_data <- left_join(arch_comp_data, AOI, by = c("studio_project_name", "trial_number", "media_name", "studio_test_name")) #All the rows were kept. No issues.

#calculate amount of trials pre-exclusions
trial_n <- arch_aoi_data %>%
  distinct(recording_name, trial_number,.keep_all = F)
#4860 trials before exclusions

```

### Create final dataset  
```{r}

#Filters participants based on predefined exclusion criteria

arch_final<- arch_aoi_data %>%
  filter(exclusion_premature=="F")%>% # we lost 16 participants and 383 trials
  filter (exclusion_low_birth_weight=="F")%>% #we lost 5 participants and 106 trials 
  filter (exclusion_health=="F")%>% # we lost 2 participants and 35 trials 
  filter (exclusion_dev_delay=="F")%>% # we lost 3 participants and 64 trials
  filter (exclusion_technical_error=="F")%>% #we lost 9 participants and 114 trials 
  filter (exclusion_testing_error=="F")%>% # we lost participants and 37 trials
  filter(other_lang_exp <= 10) #we lost 13 participants and 260 trials

#This chunk excluded 51 children

#further filters participants that have not heard French or English for at least half their lives.  

arch_final <- arch_final %>%
  mutate (half_age = age_months/2) %>%
  mutate(seq_fr_exclude = case_when(
    acq_age_fr < half_age ~ "false",
    acq_age_fr >= half_age~ "true"
  ))%>%
  mutate(seq_eng_exclude= case_when(
    acq_age_eng < half_age ~ "false",
    acq_age_eng >= half_age~ "true"
  )) %>%
  filter(seq_fr_exclude=="false")%>%
  filter(seq_eng_exclude=="false") #18 participants excluded and 340 trials

#The first dataframe keeps only the control or training trials, not the experimental condition


arch_final_fillers<- arch_final %>%
  filter(trial_type %in% c("normal", "filler", "same")) %>%
  filter(!is.na(media_name)) %>% 
  filter(is.na(key_press_event)) %>% 
  filter(is.na(studio_event))
#7 participants excluded and 1531 trials

save(arch_final_fillers, file = here("archival_KBH_anonimyzed/arch_final.Rda"))


#The second dataframe keeps only the experimental trials.
`%notin%` <- Negate(`%in%`)

arch_final_experimental<- arch_final %>%
  filter(trial_type %notin% c("normal", "filler", "same")) %>%
  filter(!is.na(media_name)) %>% 
  filter(is.na(key_press_event)) %>% 
  filter(is.na(studio_event))


save(arch_final_experimental, file = here("archival_KBH_anonimyzed/arch_final_experimental.Rda"))



```
