==> beyond_looks_analyses/1.2_prepare_KBH_archival_data_for_analysis.Rmd <==
---
title: "Untitled"
author: "Hilary Killam"
date: "2024-05-23"
output: html_document
---

```{r}

library(tidyverse)
library(here)

load(here("archival_KBH_anonimyzed/arch_final.Rda"))
load(here("archival_KBH_anonimyzed/arch_final_experimental.Rda"))

```


```{r clean filler data to be used in analyses}

kbh_data_clean <- arch_final_fillers%>%
  group_by(studio_project_name, studio_test_name, trial_number, recording_name)%>%
  mutate(trial_from_zero = recording_timestamp-min(recording_timestamp))%>%
  group_by(studio_project_name)%>%
  mutate(noun_onset = case_when(studio_project_name=="CompMix-36"~trial_from_zero-3000, #Substracting noun onset to normalize
                                studio_project_name=="LearnMix-36"~trial_from_zero-4500,
                                studio_project_name=="Mix-20"~trial_from_zero-5400,
                                studio_project_name=="Mix-14"~trial_from_zero-5400,
                                studio_project_name=="CogMisp-24"~trial_from_zero-1500))%>%
  ungroup()%>%
  rename(target_side=target, distractor_side=distractor,
         gaze_point_x= gaze_point_x_adc_spx, 
         gaze_point_y = gaze_point_y_adc_spx)%>%
  mutate(trackloss = case_when(gaze_point_x<0 ~ TRUE,
                               gaze_point_x> 1920 ~ TRUE,
                               gaze_point_y<0 ~ TRUE,
                               gaze_point_y > 1080 ~ TRUE,
                               is.na(gaze_point_x) ~ TRUE,
                               is.na(gaze_point_y) ~ TRUE,
                               is.na(validity_left) ~ TRUE,
                               is.na(validity_right) ~ TRUE,
                               validity_left > 1 ~ TRUE,
                               validity_right > 1 ~ TRUE,
                               TRUE~ FALSE)) %>% #MAKE VISUALIZATIONS ABOUT TRACKLOSS FOR SANITY CHECK
  mutate(target = case_when(gaze_point_x >= target_x_min&gaze_point_x <= target_x_max&gaze_point_y >= target_y_min&gaze_point_y <= target_y_max~TRUE, 
                            TRUE~FALSE)) %>%
  mutate(distractor = case_when(gaze_point_x >= distractor_x_min&gaze_point_x <= distractor_x_max&gaze_point_y >= distractor_y_min&gaze_point_y <= distractor_y_max~TRUE, 
                                 TRUE ~ FALSE))

#Create trial language column to describe the language in which the babies were tested
kbh_data_clean <- kbh_data_clean %>%
  ungroup()%>%
  group_by(studio_test_name, eng_exp, fre_exp)%>%
  mutate(trial_lang = case_when(str_detect(studio_test_name,"E")~ "english",
                                str_detect(studio_test_name, "F")~ "french")) %>%
  mutate(exp_to_target_lang = case_when(trial_lang == "english" ~ eng_exp,
                                        trial_lang == "french" ~ fre_exp)) %>%
  #Make trackloss rows NA for target and distractor, since we can't say for certain where the gaze is
  mutate(target = case_when(trackloss == TRUE ~ NA,
                            TRUE ~ target),
         distractor = case_when(trackloss == TRUE ~ NA,
                                TRUE ~ distractor))


#Add a column with the yoked pairs information---yoking info taken from lab server

kbh_w_items<- kbh_data_clean %>%
mutate(yoked_pair = case_when(
  grepl("Ear|Spoon", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"ear-spoon",
  grepl("Apple|Toothbrush", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"apple-toothbrush",
  grepl("Fish|Duck", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"fish-duck",
  grepl("Sheep|Monkey", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"sheep-monkey",
  grepl("Dog|Bunny", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"dog-bunny",
  grepl("Hand|Door", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"hand-door",
  grepl("Pencil|Coat", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"pencil-coat",
  grepl("Froggy|Cow", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"froggy-cow",
 
  
  grepl("Ear|Spoon", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"oreille-cuillere",
  grepl("Apple|Toothbrush", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"pomme-brosse a dents",
  grepl("Fish|Duck", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"poisson-canard",
  grepl("Sheep|Monkey", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"mouton-singe",
  grepl("Dog|Bunny", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"chien-lapin",
  grepl("Hand|Door", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"main-porte",
  grepl("Pencil|Coat", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"crayon-manteau",
  grepl("Froggy|Cow", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"grenouille-vache",
  
   
   studio_project_name== "LearnMix-36"& media_name %in% 
    c("FR_Single_ChienWalem_R.wmv", "FR_Single_LapinWalem_L.wmv", "FR_Single_LapinTeelo_R.wmv", "FR_Single_ChienTeelo_L.wmv")~ "chien-lapin",
  
     studio_project_name== "LearnMix-36"& media_name %in% 
    c("EN_Single_DogWalem_R.wmv", "EN_Single_BunnyWalem_L.wmv","EN_Single_DogTeelo_L.wmv", "EN_Single_BunnyTeelo_R.wmv")
  ~ "dog-bunny",
  
    studio_project_name== "LearnMix-36"& media_name %in% 
    c("FR_Single_LapinTeelo_L.wmv", "FR_Single_PoissonTeelo_L.wmv","FR_Single_PoissonWalem_R.wmv", "FR_Single_LapinWalem_R.wmv")
  ~ "poisson-lapin",
  
      studio_project_name== "LearnMix-36"& media_name %in% 
    c("EN_Single_BunnyTeelo_L.wmv", "EN_Single_FishTeelo_L.wmv","EN_Single_FishWalem_R.wmv", "EN_Single_BunnyWalem_R.wmv")
  ~ "fish-bunny",
  
      studio_project_name== "LearnMix-36"& media_name %in% c( "FR_Single_ChienWalem_L.wmv","FR_Single_PoissonWalem_L.wmv","FR_Single_PoissonTeelo_R.wmv", "FR_Single_ChienTeelo_R.wmv") 
  ~ "poisson-chien",
 
      studio_project_name== "LearnMix-36"& media_name %in% c( "EN_Single_FishWalem_L.wmv",
"EN_Single_DogWalem_L.wmv","EN_Single_FishTeelo_R.wmv", "EN_Single_DogTeelo_R.wmv") 
  ~ "fish-dog", 

  grepl("banC|girC", media_name) & studio_project_name== "CogMisp-24"  ~"banana-giraffe",
  grepl("monC|butC", media_name) & studio_project_name=="CogMisp-24" ~"monkey-butterfly",
  grepl("winC|mouC", media_name) & studio_project_name== "CogMisp-24" ~"window-mouth",
  grepl("pizC|tabC", media_name) & studio_project_name== "CogMisp-24" ~"pizza-table",
  grepl("choC|bowC", media_name) & studio_project_name== "CogMisp-24" ~"chocolate-bowl",
  grepl("fooC|cooC", media_name) & studio_project_name== "CogMisp-24"~"cookie-foot",


  grepl("Toothbrush|Apple", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"toothbrush-apple",
  grepl("Ear|Spoon", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"ear-spoon",
  grepl("Mouth|Door", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"mouth-door",
  grepl("Bird|Hand", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"bird-hand",
  grepl("Dog|Book", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"dog-book",
  grepl("Cookie|Foot", media_name) & studio_project_name %in% c("Mix-20","Mix-14") ~"cookie-foot",

  grepl("Brosse|Pomme", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"pomme-brosse a dents",
  grepl("Oreille|Cuillere", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"oreille-cuillere",
  grepl("Bouche|Porte", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"bouche-porte",
  grepl("Hand|Main", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"oisseau-main",
  grepl("Chien|Livre", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"chien-livre",
  grepl("Biscuit|Pied", media_name) & studio_project_name %in% c("Mix-20","Mix-14") ~"biscuit-pied",

  TRUE~"OTHER" #Learnmix doesn't have yoked pairs, so they will get OTHER here
  )
)



#Adding a column with word tested information
kbh_w_items<- kbh_w_items %>%
  mutate(target_word= case_when(
  grepl("monC", media_name) ~"monkey",
  grepl("butC", media_name) ~"butterfly",
  grepl("Door", media_name) & grepl("Mix", studio_project_name) & grepl("Eng|ENG", media_name) ~"door",
  grepl("banC", media_name) ~"banana",
  grepl("winC", media_name) ~"window",
  grepl("pizC", media_name) ~"pizza",
  grepl("fooC", media_name) ~"foot",
  grepl("bowC", media_name) ~"bowl",
  grepl("choC", media_name) ~"chocolate",
  grepl("girC", media_name) ~"giraffe",
  grepl("tabC", media_name) ~"table",
  
  
  grepl("Oreille", media_name)  & grepl("Mix", studio_project_name) ~"oreille",
  grepl("Cuillere", media_name)  & grepl("Mix", studio_project_name) ~"cuillere",
  grepl("Chien", media_name)  & grepl("Mix", studio_project_name) ~"chien",
  grepl("Biscuit", media_name)  & grepl("Mix", studio_project_name) ~"biscuit",
  grepl("Bouche", media_name)  & grepl("Mix", studio_project_name)  ~"bouche",
  grepl("Livre", media_name) & grepl("Mix", studio_project_name) ~"livre",
  grepl("Pied", media_name)  & grepl("Mix", studio_project_name) ~"pied",
  grepl("Oisseau", media_name)  & grepl("Mix", studio_project_name) ~"oisseau",
  grepl("Main", media_name)  & grepl("Mix", studio_project_name) ~"main",
  grepl("Porte", media_name)  & grepl("Mix", studio_project_name) ~"porte",
  grepl("Brosse", media_name)  & grepl("Mix", studio_project_name) ~"brosse a dents",
  grepl("Pomme", media_name)  & grepl("Mix", studio_project_name) ~"pomme",
  grepl("Poisson", media_name)  & grepl("Mix", studio_project_name) ~"poisson",
  grepl("Lapin", media_name)  & grepl("Mix", studio_project_name) ~"lapin",
  
  grepl("Ear_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "oreille",
  grepl("Spoon_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "cuillere",
  grepl("Apple_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "pomme",
  grepl("Toothbrush_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "brosse a dents",
   grepl("Fish_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "poisson",
   grepl("Duck_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "canard",
   grepl("Monkey_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "singe",
   grepl("Sheep_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "mouton",
   grepl("Dog_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "chien",
   grepl("Book_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "livre",
   grepl("Cookie_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "biscuit",
   grepl("Foot_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "pied",
   grepl("Hand_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "main",
   grepl("Door_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "porte",
   grepl("Pencil_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "crayon",
   grepl("Coat_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "manteau",
   grepl("Froggy_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "grenouille",
   grepl("Cow_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "vache",
    grepl("Bunny_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "lapin",
  
    grepl("Ear", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "ear",
  grepl("Spoon", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "spoon",
  grepl("Apple", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "apple",
  grepl("Toothbrush", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "toothbrush",
   grepl("Fish", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "fish",
   grepl("Duck", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "duck",
   grepl("Monkey", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "monkey",
   grepl("Sheep", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "sheep",
   grepl("Dog", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "dog",
   grepl("Book", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "book",
   grepl("Cookie", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "cookie",
   grepl("Foot", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "foot",
   grepl("Hand", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "hand",
   grepl("Door", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "door",
   grepl("Pencil", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "pencil",
   grepl("Coat", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "coat",
   grepl("Froggy", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "froggy",
   grepl("Cow", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "cow",
  
    grepl("Ear", media_name) & grepl("Mix", studio_project_name) ~"ear",
  grepl("Spoon", media_name)  & grepl("Mix", studio_project_name) ~"spoon",
  grepl("Dog", media_name)  & grepl("Mix", studio_project_name) ~"dog",
  grepl("Cookie", media_name)  & grepl("Mix", studio_project_name)~"cookie",
  grepl("cooC", media_name) ~"cookie",  
  grepl("Mouth", media_name)  & grepl("Mix", studio_project_name)~"mouth",
  grepl("mouC", media_name) ~"mouth",
  grepl("Book", media_name)  & grepl("Mix", studio_project_name) ~"book",
  grepl("Foot", media_name)  & grepl("Mix", studio_project_name) ~"foot",
   grepl("fooC", media_name) ~"foot",
  grepl("Bird", media_name)  & grepl("Mix", studio_project_name) ~"bird",
  grepl("Hand", media_name)  & grepl("Mix", studio_project_name) ~"hand",
  grepl("Toothbrush", media_name)   & grepl("Mix", studio_project_name) ~"toothbrush",
  grepl("Apple", media_name)  & grepl("Mix", studio_project_name) ~"apple",
  grepl("Fish", media_name)  & grepl("Mix", studio_project_name) ~"fish",
  grepl("Bunny", media_name)  & grepl("Mix", studio_project_name) ~"bunny",
  
     studio_project_name== "LearnMix-36"& media_name %in% 
    c("FR_Single_ChienWalem_R.wmv", "FR_Single_ChienTeelo_L.wmv", "FR_Single_ChienWalem_L.wmv","FR_Single_ChienTeelo_R.wmv")~ "chien",
  
       studio_project_name== "LearnMix-36"& media_name %in% 
    c("FR_Single_LapinWalem_L.wmv", "FR_Single_LapinTeelo_R.wmv", "FR_Single_LapinTeelo_L.wmv","FR_Single_LapinWalem_R.wmv")~ "lapin",
  
     studio_project_name== "LearnMix-36"& media_name %in% 
    c("EN_Single_DogWalem_R.wmv","EN_Single_DogTeelo_L.wmv", "EN_Single_DogWalem_L.wmv","EN_Single_DogTeelo_R.wmv")
  ~ "dog",
  
       studio_project_name== "LearnMix-36"& media_name %in% 
    c("EN_Single_BunnyWalem_L.wmv", "EN_Single_BunnyTeelo_R.wmv", "EN_Single_BunnyTeelo_L.wmv", "EN_Single_BunnyWalem_R.wmv")
  ~ "bunny",
  
    studio_project_name== "LearnMix-36"& media_name %in% 
    c("FR_Single_PoissonTeelo_L.wmv","FR_Single_PoissonWalem_R.wmv", "FR_Single_PoissonWalem_L.wmv","FR_Single_PoissonTeelo_R.wmv")
  ~ "poisson",
  
      studio_project_name== "LearnMix-36"& media_name %in% 
    c("EN_Single_FishTeelo_L.wmv","EN_Single_FishWalem_R.wmv","EN_Single_FishWalem_L.wmv",
"EN_Single_FishTeelo_R.wmv")
  ~ "fish",
  
  TRUE~"OTHER"
  ),
distractor_word = str_remove(yoked_pair, target_word),
distractor_word = str_remove(distractor_word, "-")
)

kbh_w_items <- kbh_w_items %>% ungroup()

save(kbh_w_items, file = here("archival_KBH_anonimyzed/kbh_w_items.Rda"))

```



```{r clean experimental data to be used in analyses}

kbh_experimental_clean <- arch_final_experimental %>%
  group_by(studio_project_name, studio_test_name, trial_number, recording_name)%>%
  mutate(trial_from_zero = recording_timestamp-min(recording_timestamp))%>%
  group_by(studio_project_name)%>%
  mutate(noun_onset = case_when(studio_project_name=="CompMix-36"~trial_from_zero-3000, #Substracting noun onset to normalize
                                studio_project_name=="LearnMix-36"~trial_from_zero-4500,
                                studio_project_name=="Mix-20"~trial_from_zero-5400,
                                studio_project_name=="Mix-14"~trial_from_zero-5400,
                                studio_project_name=="CogMisp-24"~trial_from_zero-1500))%>%
  ungroup()%>%
  rename(target_side=target, distractor_side=distractor,
         gaze_point_x= gaze_point_x_adc_spx, 
         gaze_point_y = gaze_point_y_adc_spx)%>%
  mutate(trackloss = case_when(gaze_point_x<0 ~ TRUE,
                               gaze_point_x> 1920 ~ TRUE,
                               gaze_point_y<0 ~ TRUE,
                               gaze_point_y > 1080 ~ TRUE,
                               is.na(gaze_point_x) ~ TRUE,
                               is.na(gaze_point_y) ~ TRUE,
                               is.na(validity_left) ~ TRUE,
                               is.na(validity_right) ~ TRUE,
                               validity_left > 1 ~ TRUE,
                               validity_right > 1 ~ TRUE,
                               TRUE~ FALSE)) %>% #MAKE VISUALIZATIONS ABOUT TRACKLOSS FOR SANITY CHECK
  mutate(target = case_when(gaze_point_x >= target_x_min&gaze_point_x <= target_x_max&gaze_point_y >= target_y_min&gaze_point_y <= target_y_max~TRUE, 
                            TRUE~FALSE)) %>%
  mutate(distractor = case_when(gaze_point_x >= distractor_x_min&gaze_point_x <= distractor_x_max&gaze_point_y >= distractor_y_min&gaze_point_y <= distractor_y_max~TRUE, 
                                 TRUE ~ FALSE))

#Create trial language column to describe the language in which the babies were tested
kbh_experimental_clean <- kbh_experimental_clean %>%
  ungroup()%>%
  group_by(studio_test_name, eng_exp, fre_exp)%>%
  mutate(trial_lang = case_when(str_detect(studio_test_name,"E")~ "english",
                                str_detect(studio_test_name, "F")~ "french")) %>%
  mutate(exp_to_target_lang = case_when(trial_lang == "english" ~ eng_exp,
                                        trial_lang == "french" ~ fre_exp)) %>%
  #Make trackloss rows NA for target and distractor, since we can't say for certain where the gaze is
  mutate(target = case_when(trackloss == TRUE ~ NA,
                            TRUE ~ target),
         distractor = case_when(trackloss == TRUE ~ NA,
                                TRUE ~ distractor)) %>%
  ungroup()


#Add a column with the yoked pairs information---yoking info taken from lab server

`%notin%` <- Negate(`%in%`)

kbh_experimental_items<- kbh_experimental_clean %>%
mutate(yoked_pair = case_when(
  grepl("Monkey|Sheep", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"monkey-sheep",
  grepl("Bunny|Dog", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"bunny-dog",
  grepl("Cow|Froggy", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"cow-froggy",
  grepl("Fish|Duck", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~"fish-duck",
  
  grepl("Fish|Duck", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"poisson-canard",
  grepl("Monkey|Sheep", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"mouton-singe",
  grepl("Bunny|Dog", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"chien-lapin",
  grepl("Cow|Froggy", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~"grenouille-vache",
  
  grepl("banM|girM", media_name) & studio_project_name== "CogMisp-24"  ~"banana-giraffe",
  grepl("monM|butM", media_name) & studio_project_name=="CogMisp-24" ~"monkey-butterfly",
  grepl("winM|mouM", media_name) & studio_project_name== "CogMisp-24" ~"window-mouth",
  grepl("pizM|tabM", media_name) & studio_project_name== "CogMisp-24" ~"pizza-table",
  grepl("choM|bowM", media_name) & studio_project_name== "CogMisp-24" ~"chocolate-bowl",
  grepl("fooM|cooM", media_name) & studio_project_name== "CogMisp-24"~"cookie-foot",


  grepl("Toothbrush|Apple", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"toothbrush-apple",
  grepl("Ear|Spoon", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"ear-spoon",
  grepl("Mouth|Door", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"mouth-door",
  grepl("Bird|Hand", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"bird-hand",
  grepl("Dog|Book", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"dog-book",
  grepl("Cookie|Foot", media_name) & studio_project_name %in% c("Mix-20","Mix-14") ~"cookie-foot",

  grepl("Brosse|Pomme", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"pomme-brosse a dents",
  grepl("Oreille|Cuillere", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"oreille-cuillere",
  grepl("Bouche|Porte", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"bouche-porte",
  grepl("Hand|Main", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"oisseau-main",
  grepl("Chien|Livre", media_name) & studio_project_name %in% c("Mix-20","Mix-14")  ~"chien-livre",
  grepl("Biscuit|Pied", media_name) & studio_project_name %in% c("Mix-20","Mix-14") ~"biscuit-pied",

  TRUE~"OTHER"
  )
)



#Adding a column with word tested information
kbh_experimental_items<- kbh_experimental_items %>%
  mutate(target_word= case_when(
  grepl("monM", media_name) ~"monkey",
  grepl("butM", media_name) ~"butterfly",
  grepl("Door", media_name) & grepl("Mix", studio_project_name) & grepl("Eng|ENG", media_name) ~"door",
  grepl("banM", media_name) ~"banana",
  grepl("winM", media_name) ~"window",
  grepl("pizM", media_name) ~"pizza",
  grepl("fooM", media_name) ~"foot",
  grepl("bowM", media_name) ~"bowl",
  grepl("choM", media_name) ~"chocolate",
  grepl("girM", media_name) ~"giraffe",
  grepl("tabM", media_name) ~"table",
  
  
  grepl("Oreille", media_name)  & grepl("Mix", studio_project_name) ~"oreille",
  grepl("Cuillere", media_name)  & grepl("Mix", studio_project_name) ~"cuillere",
  grepl("Chien", media_name)  & grepl("Mix", studio_project_name) ~"chien",
  grepl("Biscuit", media_name)  & grepl("Mix", studio_project_name) ~"biscuit",
  grepl("Bouche", media_name)  & grepl("Mix", studio_project_name)  ~"bouche",
  grepl("Livre", media_name) & grepl("Mix", studio_project_name) ~"livre",
  grepl("Pied", media_name)  & grepl("Mix", studio_project_name) ~"pied",
  grepl("Oisseau", media_name)  & grepl("Mix", studio_project_name) ~"oisseau",
  grepl("Main", media_name)  & grepl("Mix", studio_project_name) ~"main",
  grepl("Porte", media_name)  & grepl("Mix", studio_project_name) ~"porte",
  grepl("Brosse", media_name)  & grepl("Mix", studio_project_name) ~"brosse a dents",
  grepl("Pomme", media_name)  & grepl("Mix", studio_project_name) ~"pomme",
  grepl("Poisson", media_name)  & grepl("Mix", studio_project_name) ~"poisson",
  grepl("Lapin", media_name)  & grepl("Mix", studio_project_name) ~"lapin",

   grepl("Fish_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "poisson",
   grepl("Duck_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "canard",
   grepl("Monkey_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "singe",
   grepl("Sheep_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "mouton",
   grepl("Dog_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "chien",
   grepl("Froggy_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "grenouille",
   grepl("Cow_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "vache",
    grepl("Bunny_Fr", media_name) & studio_project_name== "CompMix-36" & grepl("F", studio_test_name) ~ "lapin",
  
   grepl("Fish", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "fish",
   grepl("Duck", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "duck",
   grepl("Monkey", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "monkey",
   grepl("Sheep", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "sheep",
   grepl("Dog", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "dog",
   grepl("Froggy", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "froggy",
   grepl("Cow", media_name) & studio_project_name== "CompMix-36" & grepl("E", studio_test_name) ~ "cow",
  
    grepl("Ear", media_name) & grepl("Mix", studio_project_name) ~"ear",
  grepl("Spoon", media_name)  & grepl("Mix", studio_project_name) ~"spoon",
  grepl("Dog", media_name)  & grepl("Mix", studio_project_name) ~"dog",
  grepl("Cookie", media_name)  & grepl("Mix", studio_project_name)~"cookie",
  grepl("cooM", media_name) ~"cookie",  
  grepl("Mouth", media_name)  & grepl("Mix", studio_project_name)~"mouth",
  grepl("mouM", media_name) ~"mouth",
  grepl("Book", media_name)  & grepl("Mix", studio_project_name) ~"book",
  grepl("Foot", media_name)  & grepl("Mix", studio_project_name) ~"foot",
   grepl("fooM", media_name) ~"foot",
  grepl("Bird", media_name)  & grepl("Mix", studio_project_name) ~"bird",
  grepl("Hand", media_name)  & grepl("Mix", studio_project_name) ~"hand",
  grepl("Toothbrush", media_name)   & grepl("Mix", studio_project_name) ~"toothbrush",
  grepl("Apple", media_name)  & grepl("Mix", studio_project_name) ~"apple",
  grepl("Fish", media_name)  & grepl("Mix", studio_project_name) ~"fish",
  grepl("Bunny", media_name)  & grepl("Mix", studio_project_name) ~"bunny",
  
  TRUE~"OTHER"
  ),
distractor_word = str_remove(yoked_pair, target_word),
distractor_word = str_remove(distractor_word, "-")
)

kbh_experimental_items <- kbh_experimental_items %>% ungroup()



save(kbh_experimental_items, file = here("archival_KBH_anonimyzed/kbh_experimental_items.Rda"))


```
