==> beyond_looks_analyses/3_loading_cleaning_pkbank.Rmd <==
---
title: "Peekbank"
output: html_document
date: "2024-10-23"
---

```{r setup, include=FALSE}

#remotes::install_github("peekbank/peekbankr")
library(peekbankr)
library(tidyverse)
library(dbplyr)
library(tidylog)
library(here)

`%notin%` <- negate(`%in%`)

```

#Get data
```{r}

#Get peekbank data - start with 5 datasets that have been code checked (we are NOT using fmw_2013 anymore because of data insonsistency issues)


code_checked_studies <- c("casillas_tseltal_2015", "ronfard_2021", "pomper_saffran_2016", "weisleder_stl", "xsectional_2007", "swingley_aslin_2002", "potter_remix", "hurtado_2008", "garrison_bergelson_2020", "bacon_gendercues", "frank_tablet_2016", "baumgartner_2014")

peekbank_data <- get_aoi_timepoints(dataset_name = code_checked_studies) %>%
  left_join(get_administrations(dataset_name = code_checked_studies)) %>%
  left_join(get_subjects() %>% select(subject_id, sex)) %>%
  left_join(get_trials(dataset_name = code_checked_studies)) %>%
  left_join(get_trial_types(dataset_name = code_checked_studies)) %>%
  left_join(get_stimuli(dataset_name = code_checked_studies) %>% rename(target_id = stimulus_id)) %>%
  left_join(get_stimuli(dataset_name = code_checked_studies) %>% rename(distractor_id = stimulus_id) %>%
              dplyr::select(distractor_stimulus_label = english_stimulus_label, distractor_id, dataset_id, dataset_name)) %>% 
  group_by(dataset_id, dataset_name, administration_id) %>%
  #create numeric trial number that starts at 1
  mutate(trial_number = trial_id - min(trial_id) + 1,
         trial_index = data.table::rleid(trial_number)) %>%
  ungroup() %>%
  mutate(exp_type = case_when(condition == "familiar-familiar" ~ "non-experimental",
                              condition == "familiar-NOVEL" ~ "experimental",
                              condition == "FAMILIAR-novel" ~ "experimental",
                              is.na(condition) & dataset_name == "casillas_tseltal_2015" ~ "non-experimental",
                              condition == "Unique" ~ "experimental",
                              condition == "Generic" ~ "non-experimental",
                              condition == "color" ~ "experimental",
                              condition == "target_image" ~ "non-experimental",
                              condition == "cp" ~ "non-experimental",
                              condition == "m-e" ~ "experimental",
                              condition == "m-h" ~ "experimental",
                              condition == "filler" ~ "DELETE",
                              str_detect(condition, "^mix_")  ~ "experimental",
                              str_detect(condition, "^same_") ~ "non-experimental",
                              #is.na(condition) & dataset_name == "fmw_2013" ~ "non-experimental",
                              is.na(condition) & dataset_name == "hurtado_2008" ~ "non-experimental",
                              is.na(condition) & dataset_name == "ronfard_2021" ~ "non-experimental",
                              is.na(condition) & dataset_name == "weisleder_stl" ~ "non-experimental",
                              is.na(condition) & dataset_name == "xsectional_2007" ~ "non-experimental")) %>%
  #rename columns to match our own dataset column names
  rename(age_months = age,
         target_word = original_stimulus_label,
         noun_onset = t_norm,
         media_name = english_stimulus_label)  %>%
  mutate(target = case_when(aoi == "target" ~ 1,
                            aoi == "missing" ~ NA_real_, 
                            TRUE ~ 0),
         distractor = case_when(aoi == "distractor" ~ 1,
                            aoi == "missing" ~ NA_real_, 
                            TRUE ~ 0),
         trackloss = case_when(aoi == "missing" ~ TRUE,
                            TRUE ~ FALSE),
         age_months = floor(age_months), #fix a few months that have decimal points, since none of the other ones do
         recording_name = paste0(subject_id, "_", administration_id, "_", dataset_name, "_", age_months)) %>%
  filter(!is.na(age_months)) %>% #one kid has NA for age and it's messing things up for the ICC calculation... not sure we can do any real analysis for an unknown age, so remove them.
  filter(exp_type != "DELETE") %>%
      #one Peekbank dataset has duplicate rows for some reason, so fix this
      distinct()

#add the same data to itself with the exp_type "both_types" for new analyses that aren't split between types

peekbank_data <- peekbank_data %>%
  bind_rows(peekbank_data %>% mutate(exp_type = "both_types")) 

beepr::beep()

```


#Clean data, trackloss exclusions

```{r, message=FALSE}

print_list <- list()

#-----------------------------------------Exclusions based on looking data (before segmenting)

####### Only keep kids with at least 50% of interest period looking AND who have at least 25% of trials
## Total trials for each study is usually 12 (CogMisp, Mix), except CompMix, CompLearn with 16, LearnMix with 6

#Some peekbank datasets have multiple ages where each age was shown a different number of trials. To correctly calculate trial cutoff, need to know the intended study age and max number of trials.

max_trials <- read_csv(here("max_trials_peekbank.csv"))

trial_cutoff <- peekbank_data %>%
  filter(exp_type == "both_types") %>%
  group_by(recording_name, dataset_name, age_months) %>%
  distinct(trial_number) %>%
  count(recording_name) %>%
  rename(n_trials_subject = n) %>%
  mutate(intended_study_age = case_when(dataset_name %in% c("casillas_tseltal_2015") ~ 0,
                                        dataset_name %in% c("frank_tablet_2016", "xsectional_2007") ~ 12,
                                        dataset_name %in% c("swingley_aslin_2002") ~ 14,
                                        dataset_name %in% c("potter_remix", "ronfard_2021") ~ 18,
                                        dataset_name %in% c("weisleder_stl") & age_months < 24 ~ 18,
                                        dataset_name %in% c("fmw_2013", "hurtado_2008") & age_months < 22 ~ 18,
                                        dataset_name %in% c("fmw_2013", "hurtado_2008") & age_months >= 22 ~ 24,
                                        dataset_name %in% c("garrison_bergelson_2020") ~ 24,
                                        dataset_name %in% c("pomper_saffran_2016") ~ 36
                                        )) %>%
  left_join(max_trials)



n_subjects_original <- peekbank_data %>%
  filter(exp_type == "both_types") %>%
  distinct(recording_name, dataset_name) %>%
  count(dataset_name)

n_trials_original <- peekbank_data %>%
  filter(exp_type == "both_types") %>%
  distinct(recording_name, dataset_name, trial_number) %>%
  nrow()

print_list[[paste("Pre-exclusions N subjects:")]] <- n_subjects_original
print_list[[paste("Pre-exclusions N trials:")]] <- n_trials_original
  
trackloss_cleaned <- peekbank_data %>%
  filter(exp_type == "both_types") %>%
  #filter to interest period
  filter(noun_onset>= 360 & noun_onset <= 3000) %>%
  #for each kid in each trial,
  group_by(recording_name, trial_number) %>%
  #get trackloss percentage
  mutate(trackloss_pct = sum(aoi == "missing")/n(),
         trackloss_cutoff = case_when(trackloss_pct <.5 ~ "keep",
                                      trackloss_pct >= .5 ~ "exclude")) %>%
  ungroup() %>%
  distinct(recording_name, trial_number, trackloss_pct, trackloss_cutoff, dataset_name) %>%
  filter(trackloss_cutoff == "keep") %>%
  group_by(recording_name) %>%
  mutate(n_valid_trials_subject = n()) %>%
  ungroup() %>%
  dplyr::select(dataset_name, recording_name, trial_number, trackloss_pct, n_valid_trials_subject)

n_subjects_after_trackloss_exclusion <- trackloss_cleaned %>%
  distinct(recording_name, dataset_name) %>%
  count(dataset_name)

n_trials_after_trackloss_exclusion <- trackloss_cleaned %>%
  distinct(recording_name, dataset_name, trial_number) %>%
  nrow()

print_list[[paste("After Trackloss Cleaning N subjects:")]] <- n_subjects_after_trackloss_exclusion
print_list[[paste("After Trackloss Cleaning N trials:")]] <- n_trials_after_trackloss_exclusion
  
peekbank_data_after_exclusions <- peekbank_data %>%
  right_join(trackloss_cleaned) %>%
  left_join(trial_cutoff) %>%
  mutate(trial_n_cutoff = round(max_trials*.25, digits = 0)) %>%
  filter(n_valid_trials_subject >= trial_n_cutoff) %>%
  ungroup()

n_subjects_after_trackloss_and_trials_exclusion <- peekbank_data_after_exclusions %>%
  filter(exp_type == "both_types") %>%
  distinct(recording_name, dataset_name) %>%
  count(dataset_name) #XX kids lost from these exclusions

n_trials_after_trackloss_and_trials_exclusion <- peekbank_data_after_exclusions %>%
  filter(exp_type == "both_types") %>%
  distinct(recording_name, dataset_name, trial_number) %>%
  nrow() #XX kids lost from these exclusions

print_list[[paste("After Trackloss and Trials Exclusions N subjects:")]] <- n_subjects_after_trackloss_and_trials_exclusion
print_list[[paste("After Trackloss and Trials Exclusions N trials:")]] <- n_trials_after_trackloss_and_trials_exclusion

```

#Create yoked pairs

```{r}
#these studies appear to have yoked pairs: 

yoked_studies <- c("hurtado_2008", "pomper_saffran_2016", "potter_remix", "ronfard_2021", "swingley_aslin_2002", "weisleder_stl") 

peekbank_data_yoked_pairs <- peekbank_data_after_exclusions %>%
  filter(dataset_name %in% yoked_studies) %>% 
  mutate(yoked_target = case_when(target_word %in% c("shawl", "gall", "pelota") ~ "ball",
                                     target_word %in% c("mog", "tog", "perro") ~ "dog",
                                     target_word %in% c("opal", "opple", "manzana") ~ "apple",
                                     target_word %in% c("pity", "yitty", "gato") ~ "kitty",
                                     target_word %in% c("raby", "vaby") ~ "baby",
                                     target_word %in% c("cur", "kier") ~ "car",
                                     target_word %in% c("plátano") ~ "banana",
                                     target_word %in% c("leche") ~ "milk",
                                     target_word %in% c("puerta") ~ "door",
                                     target_word %in% c("boca") ~ "mouth",
                                     target_word %in% c("globo") ~ "balloon",
                                     target_word %in% c("galleta") ~ "cookie",
                                     target_word %in% c("pie") ~ "foot",
                                     target_word %in% c("jugo") ~ "juice",
                                     target_word %in% c("libro") ~ "book",
                                     target_word %in% c("cuchara") ~ "spoon",
                                     target_word %in% c("zapato") ~ "shoe",
                                     target_word %in% c("caballo") ~ "horse",
                                    target_word %in% c("pájaro") ~ "bird",
                                    TRUE ~ target_word),
         yoked_distractor = case_when(distractor_stimulus_label %in% c("shawl", "gall", "pelota") ~ "ball",
                                     distractor_stimulus_label %in% c("mog", "tog", "perro") ~ "dog",
                                     distractor_stimulus_label %in% c("opal", "opple", "manzana") ~ "apple",
                                     distractor_stimulus_label %in% c("pity", "yitty", "gato") ~ "kitty",
                                     distractor_stimulus_label %in% c("raby", "vaby") ~ "baby",
                                     distractor_stimulus_label %in% c("cur", "kier") ~ "car",
                                     distractor_stimulus_label %in% c("plátano") ~ "banana",
                                     distractor_stimulus_label %in% c("leche") ~ "milk",
                                     distractor_stimulus_label %in% c("puerta") ~ "door",
                                     distractor_stimulus_label %in% c("boca") ~ "mouth",
                                     distractor_stimulus_label %in% c("globo") ~ "balloon",
                                     distractor_stimulus_label %in% c("galleta") ~ "cookie",
                                     distractor_stimulus_label %in% c("pie") ~ "foot",
                                     distractor_stimulus_label %in% c("jugo") ~ "juice",
                                     distractor_stimulus_label %in% c("libro") ~ "book",
                                     distractor_stimulus_label %in% c("cuchara") ~ "spoon",
                                     distractor_stimulus_label %in% c("zapato") ~ "shoe",
                                     distractor_stimulus_label %in% c("caballo") ~ "horse",
                                    distractor_stimulus_label %in% c("pájaro") ~ "bird",
                                    TRUE ~ distractor_stimulus_label),
         yoked_pair = case_when(yoked_target < yoked_distractor ~ paste0(yoked_target, "-", yoked_distractor),
                                TRUE ~ paste0(yoked_distractor, "-", yoked_target))) %>% 
  distinct(recording_name, subject_id, dataset_name, trial_number, age_months, yoked_target, yoked_distractor, yoked_pair)
  

```

#Make variables
```{r}

for (i in 1:n_distinct(peekbank_data$exp_type)) {
  
  analysis_type <- unique(peekbank_data$exp_type)[[i]]
  analysis_dataset <- peekbank_data_after_exclusions %>%
    filter(exp_type == analysis_type)

#create initial look variable 

initial_looks <- analysis_dataset %>%
  filter(noun_onset >= 260 & noun_onset < 360) %>%
  group_by(recording_name, trial_number, media_name) %>% 
     summarise(
    distractor_count = sum(aoi == "distractor"),
    target_count = sum(aoi == "target"),
    trackloss_count = sum(aoi == "missing") + sum(is.na(aoi))
     )%>%
  #Changing how this is done
  # mutate(initial_look = case_when(target_count >0 ~ "target", #if the child *ever* looks to target during first 100 ms, then counted as initial target look
  #                                 distractor_count/(distractor_count+trackloss_count) >=0.5 ~ "distractor",
  #                                 TRUE ~ "neither")) %>%
  mutate(initial_look = case_when(target_count > distractor_count ~ "target",
                                  distractor_count > target_count ~ "distractor",
                                  TRUE ~ "neither")) %>%
  ungroup() 


analysis_dataset <- left_join(analysis_dataset, initial_looks, by= c("recording_name", "trial_number", "media_name"))

for (j in c("target", "distractor", "all")) {
  
  if (j %in% c("target", "distractor")) {
  
  analysis_dataset_split <- analysis_dataset%>%
    filter(!is.na(initial_look)) %>%
    filter(initial_look == j)
  } else if (j == "all") {
    # 15 trials from the garrison_bergelson_2020 dataset don't have data for the timestamps required to calculate initial_look, so exclude these ones
    analysis_dataset_split <- analysis_dataset %>%
      filter(!is.na(initial_look))
  }

#1. Proportion looking to target

prop_looking_by_trial <- analysis_dataset_split %>%
  filter(noun_onset>= 360 & noun_onset <= 3000)%>%
  group_by(recording_name, subject_id, trial_number, media_name, 
           age_months, initial_look) %>%
  summarise(samples_target=sum(target, na.rm = T),
            samples_distractor=sum(distractor, na.rm = T))%>%
  mutate(prop_looking = samples_target/(samples_target + samples_distractor)) %>%
  ungroup()

prop_looking_by_subject <- prop_looking_by_trial %>%
  group_by(recording_name) %>%
  summarise(mean_prop_looking = mean(prop_looking),
            sd_prop_looking = sd(prop_looking),
            n_trials_prop_looking = length(unique(trial_number))) %>%
  ungroup()


#2. Proportion looking difference 
#Within a pair of pictures A and B, the fixation to picture A relative to B when A was the target, minus the fixation #to A when A was the distractor.

prop_looking_diff_by_item <- analysis_dataset_split %>%
  filter(dataset_name %in% yoked_studies) %>%
  ungroup() %>%
  left_join(peekbank_data_yoked_pairs) %>%
  filter(noun_onset>= 360 & noun_onset <= 3000)%>%
  dplyr::select(recording_name, noun_onset, yoked_target, yoked_distractor, target, distractor)%>%
  pivot_longer(c(yoked_target, yoked_distractor), names_to = "word_type", values_to = "word") %>%
  mutate(looking = case_when(word_type == "yoked_target" & target == TRUE ~ TRUE,
                             word_type == "yoked_distractor" & distractor == TRUE ~ TRUE,
                             TRUE ~ FALSE)) %>%
  group_by(recording_name, word_type, word) %>%
  #what percent of the time when x is target are kids looking at x, what percent of the time when x is distractor are kids looking at x
  summarise(prop_looking = sum(looking == T, na.rm = T)/n()) %>%
  pivot_wider(names_from = "word_type", values_from = "prop_looking") %>%
  filter(!is.na(yoked_target) & !is.na(yoked_distractor)) %>%
  mutate(prop_looking_diff = yoked_target - yoked_distractor) %>%
  dplyr::select(recording_name, yoked_target = word, prop_looking_diff) %>%
  ungroup() 

prop_looking_diff_by_subject <- prop_looking_diff_by_item %>%  
  group_by(recording_name) %>% #Calculating this variable by subject instead of by trial for the correlation matrix.
  summarise(mean_prop_looking_diff = mean(prop_looking_diff),
            sd_prop_looking_diff = sd(prop_looking_diff),
            n_trials_prop_looking_diff = length(unique(yoked_target))) %>%
  ungroup() #only 165 kids have data for this variable, because they didn't have enough data/trials included with both words of each yoked pair


#3. Number of fixations

#For this one, stretches of same looks are counted as one fixation
fixations_total_by_trial <- analysis_dataset_split %>%
  ungroup() %>%
  filter(noun_onset >= 360 & noun_onset <= 3000) %>%
  group_by(recording_name, trial_number) %>%
  mutate(trackloss_group_id = data.table::rleid(trackloss)) %>%
  group_by(trackloss_group_id, recording_name, trial_number) %>%
  mutate(num_consec_trackloss_rows = n()) %>%
  ungroup() %>%
  filter(trackloss == FALSE | (trackloss == TRUE & num_consec_trackloss_rows >= 2)) %>%
  group_by(recording_name, trial_number) %>% 
  mutate(looking_stretch_id = data.table::rleid(target)) %>%
  group_by(recording_name, trial_number, looking_stretch_id) %>%
  #filter to fixations of at least 8 rows or ~ 200 ms
  filter(n() >= 8) %>%
  #filter to target fixations
  filter(target == TRUE) %>%
  distinct(recording_name, trial_number, looking_stretch_id, initial_look) %>%
  group_by(recording_name, trial_number, initial_look) %>%
  summarise(number_fixations = n()) %>%
  ungroup()

#before, this was just counting number of fixations... changed to mean number fixations per trial per child
mean_num_fixations_by_subject <- fixations_total_by_trial %>%
  group_by(recording_name) %>%
  summarise(mean_number_fixations = mean(number_fixations),
            sd_number_fixations = sd(number_fixations),
            n_trials_number_fixations = length(unique(trial_number))) %>%
  ungroup()



#3.1 Longest fixation to target

#For this one, stretches of same looks are counted as one fixation
longest_look_by_trial <- analysis_dataset_split %>%
  ungroup() %>%
  filter(noun_onset >= 360 & noun_onset <= 3000) %>%
  group_by(recording_name, trial_number) %>%
  mutate(trackloss_group_id = data.table::rleid(trackloss)) %>%
  group_by(trackloss_group_id, recording_name, trial_number) %>%
  mutate(num_consec_trackloss_rows = n()) %>%
  ungroup() %>%
  filter(trackloss == FALSE | (trackloss == TRUE & num_consec_trackloss_rows >= 2)) %>%
  group_by(recording_name, trial_number) %>% 
  mutate(looking_stretch_id = data.table::rleid(target)) %>%
  group_by(recording_name, trial_number, looking_stretch_id) %>%
  #filter to fixations of at least 8 rows or ~ 200 ms
  filter(n() >= 8) %>%
  #filter to target fixations
  filter(target == TRUE) %>%
  count(recording_name, trial_number, looking_stretch_id, initial_look, sample_rate) %>%
  group_by(recording_name, trial_number, initial_look) %>%
  filter(n == max(n)) %>%
  slice_head(n = 1) %>%
  ungroup() %>%
  mutate(longest_look_by_trial = n*(1000/sample_rate)) %>%
  dplyr::select(-n, -sample_rate, -looking_stretch_id)

#before, this was just counting number of fixations... changed to mean number fixations per trial per child
mean_longest_look_by_subject <- longest_look_by_trial %>%
  group_by(recording_name) %>%
  summarise(mean_longest_look = mean(longest_look_by_trial),
            sd_longest_look = sd(longest_look_by_trial),
            n_trials_number_fixations = length(unique(trial_number))) %>%
  ungroup() 


#4. Correct first shift to target
#On trials where infants were fixated on the distractor upon hearing the target word (i.e., distractor-initial trials), does the kid shift to the target picture within the 300–1,800 ms window following target word onset, as a proportion of all distractor-initial trials.
  

first_shift_target <- analysis_dataset_split %>%
  ungroup() %>%
  filter(noun_onset >= 360 & noun_onset <= 1800) %>% #Keeping the shorter window to be consistent with Fernald

  
    #Grouping and summarizing to get by subject by trial flagging of whether there was switching or not.
     group_by(recording_name, subject_id, trial_number, media_name, initial_look) %>%  
  summarise(target = sum(target, na.rm = T),
            distractor =  sum(distractor, na.rm = T)) %>% #Counting how many looks to target and distractor by timeslice grouped by target or distractor initial trials.
  ungroup() %>%
  mutate(correct_shift = case_when(initial_look == "target" & distractor ==0 ~ TRUE,
                                   initial_look == "distractor" & target >0 ~ TRUE,
                                   TRUE~FALSE)) #Making a column to calculate correct shifts. We defined correct shift as never looking to the distractor on target initial trial, or switching to the taget on distractor initial trials

prop_shifts_distractor_initial_by_trial <- first_shift_target %>%
    filter(initial_look == "distractor") 

prop_shifts_distractor_initial_by_subject <- prop_shifts_distractor_initial_by_trial %>%
  group_by(recording_name, initial_look) %>%
    summarise(mean_prop_shift_distractor_initial = sum(correct_shift == TRUE)/n(),
              sd_prop_shift_distractor_initial = sd(as.numeric(correct_shift)),
            n_trials_prop_shift_distractor_initial = length(unique(trial_number))) #some SDs are NA because the child only has one row, and no standard deviation can be calculated from one value
  
prop_shifts_target_initial_by_trial <- first_shift_target %>%
    filter(initial_look == "target") 

prop_shifts_target_initial_by_subject <- prop_shifts_target_initial_by_trial %>%
  group_by(recording_name, initial_look) %>%
    summarise(mean_prop_shift_target_initial = sum(correct_shift == TRUE)/n(),
              sd_prop_shift_target_initial = sd(as.numeric(correct_shift)),
            n_trials_prop_shift_target_initial = length(unique(trial_number))) #some SDs are NA because the child only has one row, and no standard deviation can be calculated from one value
  

#5. Reaction time

#do we want the latency from noun onset (0) or from window of analysis(360)? Right now using window of analysis

RT_by_trial <- analysis_dataset_split %>%
  ungroup() %>%
  filter(noun_onset >= 360 & noun_onset <= 1800) %>% #Keeping the shorter window to be consistent with Fernald
    #Grouping and summarizing to get by subject by trial flagging of whether there was switching or not.
  filter(initial_look == "distractor") %>%
  filter(target == TRUE) %>%
  group_by(trial_number, media_name, recording_name, subject_id, initial_look, target) %>%
  slice_head() %>%
  mutate(latency_to_switch = noun_onset - 360) %>%
  ungroup() %>%
  dplyr::select(recording_name, trial_number, latency_to_switch, initial_look) 

RT_by_subject <- RT_by_trial %>%
  group_by(recording_name) %>%
  summarise(mean_RT = mean(latency_to_switch, na.rm = T),
            sd_RT = sd(latency_to_switch, na.rm = T),
            n_trials_RT = length(unique(trial_number))) %>%
  ungroup()


#6. Total number of switches between AOIs

number_of_switches_by_trial <- analysis_dataset_split %>%
  ungroup() %>%
  filter(noun_onset >= 360 & noun_onset <= 3000) %>%
  #for this one, get rid of track loss rows
  filter(trackloss == FALSE) %>%
  group_by(recording_name, trial_number) %>% 
  mutate(looking_stretch_id = data.table::rleid(target)) %>%
  group_by(recording_name, trial_number, looking_stretch_id) %>%
  #keep stretches that are at least 200 ms or 8 rows
  filter(n() >= 8) %>%
  mutate(where_looking = case_when(target == TRUE ~ "target",
                                   distractor == TRUE ~ "distractor",
                                   TRUE ~ "neither")) %>%
  filter(where_looking != "neither") %>%
  distinct(recording_name, trial_number, looking_stretch_id, where_looking, media_name, subject_id, initial_look) %>%
  group_by(recording_name, trial_number) %>%
  mutate(AOI_switches = case_when(lag(where_looking) == "target" & where_looking == "distractor" ~ 1,
                            lag(where_looking) == "distractor" & where_looking == "target" ~ 1)) %>%
  group_by(recording_name, trial_number, initial_look) %>%
  summarise(total_aoi_switches = sum(AOI_switches, na.rm = T)) %>%
  ungroup()

number_of_switches_by_subject <- number_of_switches_by_trial %>%
  group_by(recording_name) %>%
  summarise(mean_aoi_switches = mean(total_aoi_switches, na.rm = T),
            n_trials_mean_aoi_switches = length(unique(trial_number))) %>%
  ungroup()


#8. Mean fixation duration during trial
#Mean duration of fixations during the window of analysis within a trial. Mean by subject by trial?

mean_fixation_dur_by_trial <- analysis_dataset_split %>%
  filter(noun_onset>= 360 & noun_onset <= 3000)%>%
  group_by(recording_name, trial_number) %>%
  mutate(trackloss_group_id = data.table::rleid(trackloss)) %>%
  group_by(trackloss_group_id, recording_name, trial_number) %>%
  mutate(num_consec_trackloss_rows = n()) %>%
  ungroup() %>%
  filter(trackloss == FALSE | (trackloss == TRUE & num_consec_trackloss_rows >= 2)) %>%
  group_by(recording_name, trial_number) %>% 
  mutate(looking_stretch_id = data.table::rleid(target)) %>%
  group_by(recording_name, trial_number, looking_stretch_id) %>%
  #filter to fixations of at least 8 rows or ~ 200 ms
  filter(n() >= 8) %>%
  filter(target == TRUE) %>%
  group_by(recording_name, trial_number, media_name, looking_stretch_id, initial_look, sample_rate) %>%
  summarise(length_of_looks = n()) %>%
  mutate(duration_of_look = length_of_looks*(1000/sample_rate)) %>%
  group_by(recording_name, trial_number, initial_look) %>%
  summarise(duration_of_look = mean(duration_of_look, na.rm = T)) %>%
  ungroup()

mean_fixation_dur_by_subject <- mean_fixation_dur_by_trial %>% 
  group_by(recording_name) %>%
  summarise(mean_duration_of_look = mean(duration_of_look, na.rm = T),
            sd_duration_of_look = sd(duration_of_look, na.rm = T),
            n_trials_duration_of_look = length(unique(trial_number))) %>%
  ungroup()


#11. Duration of first look
### Duration of the first look recorded toward a particular AOI. - To target specifically, with at least 200 ms fixation length

first_look_duration_by_trial <- analysis_dataset_split %>%
  ungroup() %>%
  filter(noun_onset >= 360 & noun_onset <= 3000) %>%
  group_by(recording_name, trial_number) %>%
  mutate(trackloss_group_id = data.table::rleid(trackloss)) %>%
  group_by(trackloss_group_id, recording_name, trial_number) %>%
  mutate(num_consec_trackloss_rows = n()) %>%
  ungroup() %>%
  filter(trackloss == FALSE | (trackloss == TRUE & num_consec_trackloss_rows >= 2)) %>%
  group_by(recording_name, trial_number) %>% 
  mutate(looking_stretch_id = data.table::rleid(target)) %>%
  group_by(recording_name, trial_number, looking_stretch_id) %>%
  mutate(row = row_number()) %>% 
  mutate(where_looking = case_when(target == TRUE ~ "target",
                                   distractor == TRUE ~ "distractor",
                                   TRUE ~ "neither")) %>%
  filter(where_looking == "target") %>%
  mutate(first_look_dur = max(row_number())*(1000/sample_rate)) %>% 
  filter(first_look_dur >= 200) %>%
  group_by(trial_number, recording_name) %>%
  filter(looking_stretch_id == min(looking_stretch_id)) %>% 
  ungroup() %>%
  distinct(recording_name, trial_number, media_name, first_look_dur, initial_look) %>%
  ungroup()

first_look_duration_by_subject <- first_look_duration_by_trial %>%
  group_by(recording_name) %>%
  summarise(mean_first_look_dur = mean(first_look_dur, na.rm = T),
            sd_first_look_dur = sd(first_look_dur, na.rm = T),
            n_trials_first_look_dur = length(unique(trial_number))) %>%
  ungroup()
 
data_new_dvs_by_trial <- analysis_dataset_split %>%
  distinct(recording_name, trial_number, subject_id, dataset_name, age_months, target_word) %>%
  left_join(peekbank_data_yoked_pairs) %>%
  left_join(prop_looking_by_trial) %>%
  #double check if kids see the same word more than once... if so, this item will be ducplicated in the dataset
  left_join(prop_looking_diff_by_item) %>%
  # commenting out for now until we verify how to do this one left_join(sampling_fixations_total %>% dplyr::select(-subject_id)) %>%
  left_join(fixations_total_by_trial) %>%
  left_join(prop_shifts_distractor_initial_by_trial %>% dplyr::select(-media_name, -target, -distractor)) %>%
  left_join(prop_shifts_target_initial_by_trial %>% dplyr::select(-media_name, -target, -distractor) %>% rename(remain_on_target = correct_shift)) %>%
  left_join(RT_by_trial) %>%
  left_join(number_of_switches_by_trial) %>%
  left_join(mean_fixation_dur_by_trial) %>%
  left_join(first_look_duration_by_trial) %>%
  left_join(longest_look_by_trial) %>%
  ungroup() %>%
  mutate(age_months_centred = age_months - 25,
         correct_shift = case_when(correct_shift == T ~ 1,
                                   correct_shift == F ~ 0),
         remain_on_target = case_when(remain_on_target == T ~ 1,
                                   remain_on_target == F ~ 0))

data_new_dvs_by_subject <- analysis_dataset_split %>%
  distinct(recording_name, subject_id, dataset_name, age_months) %>% #Added french and english exposure to be able to calculate balance for some analyses.
  left_join(prop_looking_by_subject) %>%
  #double check if kids see the same word more than once... if so, this item will be ducplicated in the dataset
  left_join(prop_looking_diff_by_subject) %>%
  # commenting out for now until we verify how to do this one left_join(sampling_fixations_total %>% dplyr::select(-subject_id)) %>%
  left_join(mean_num_fixations_by_subject) %>%
  left_join(prop_shifts_target_initial_by_subject %>% select(-initial_look)) %>%
  left_join(prop_shifts_distractor_initial_by_subject %>% select(-initial_look)) %>%
  left_join(RT_by_subject) %>%
  left_join(number_of_switches_by_subject) %>%
  left_join(mean_fixation_dur_by_subject) %>%
  left_join(first_look_duration_by_subject) %>%
  left_join(mean_longest_look_by_subject) %>%
  ungroup()%>%
  mutate(age_months_centred = age_months - 25)


#export to later get means, SDs for calculating Cohen's D when dataset is run with the other experimental type
data_new_dvs_by_subject %>%
  mutate(data_split = j) %>%
  write_csv(here(paste0("all_variables_datasets/peekbank_by_subject_means_sds_", analysis_type, "_", j, ".csv")))

data_new_dvs_by_trial %>%
  mutate(data_split = j) %>% 
  write_csv(here(paste0("all_variables_datasets/peekbank_by_trial_means_sds_", analysis_type, "_", j, ".csv")))


#How many trials in average when we collapse by subject?

num_trials_per_subject <- round(data_new_dvs_by_trial %>%
  group_by(recording_name) %>%
    summarise(num_trials = length(unique(trial_number))) %>%
    summarise(avg_num_trials = mean(num_trials)) %>% pull(avg_num_trials), digits = 1)

print_list[[paste(analysis_type, "average number of trials per participant (", j, "):")]] <- num_trials_per_subject

#Save datasets to load and use in next section where non-exp and exp are both needed
# 
# save(data_new_dvs_by_subject, file = here(paste0("all_variables_datasets/peekbank_bysubject_", analysis_type, "_", j, ".Rda")))

}
}

```



```{r}
#Save the print_list to a text file

#sink(here(paste0("beyond_looks_analyses/", analysis_type, "_code_printout.txt"))); print(print_list); sink()

```
